Organize file structure and syntax
